Develop

+++ system0/html/php/login/v3/php/cloud.php
@@ -52,8 +52,8 @@ function load_user()
 	function get_base64_preview($filename){
 		$base64="";
 		$file=fopen($filename,"r");
-		$start=0;
-		while(!feof($file)){
+		$start=-1;
+		while(!feof($file)&&$start!=0){
 			$buf=fgets($file);
 			if(stripos($buf,"thumbnail end")!==false)
 				$start=0;
@@ -85,11 +85,15 @@ function load_user()
   
 </head>
 <body>
-	<div class="container mt-10" style="height: 100vh;overflow-y:auto">
+	<div class="container mt-4" style="height: 100vh;overflow-y:auto">
 		<div class="row justify-content-center">
 			<!--<div style="width: 90vh">-->
 			      <h1>Deine Dateien</h1>
 				<div class="container">
+					<form action="cloud.php" method="POST">
+						<input type="text" name="search" placeholder="Suchbegriff">
+						<button type="submit" class="btn btn-dark my-5">Suchen</button>
+					</form>
 				  <table class="table">
 				    <thead>
 				      <tr>
@@ -112,14 +116,27 @@ function load_user()
 					  // Iterate through the files and display them in the table
 					  $count = 1;
 					  foreach ($files as $file) {
-					      echo '<tr>';
-					      echo '<td><img  style="display:block; width:100px;height:100px;" id="base64image" src="data:image;base64,' . get_base64_preview($file) . '"/></td>';
-					      echo '<td>' . basename($file) . '</td>';
-					      echo '<td><a href="print.php?cloudprint='.basename($file).'">Drucken</a></td>';
-					      echo "<td><a href='cloud.php?delete=".basename($file)."' >" . "Löschen" . '</a></td>';
-					      echo "<td><a href='/system0/html/user_files/$username/".basename($file)."' download>" . "Herunterladen" . '</a></td>';
-					      echo "<td><a href='cloud.php?public=".basename($file)."'>Öffentlich verfügbar machen</a></td>";
-					      echo '</tr>';
+						if(isset($_POST["search"])){
+							if (stripos(basename($file), $_POST["search"]) !== false) {
+							      echo '<tr>';
+							      echo '<td><img  style="display:block; width:100px;height:100px;" id="base64image" src="data:image;base64,' . get_base64_preview($file) . '"/></td>';
+							      echo '<td>' . basename($file) . '</td>';
+							      echo '<td><a href="print.php?cloudprint='.basename($file).'">Drucken</a></td>';
+							      echo "<td><a href='cloud.php?delete=".basename($file)."' >" . "Löschen" . '</a></td>';
+							      echo "<td><a href='/system0/html/user_files/$username/".basename($file)."' download>" . "Herunterladen" . '</a></td>';
+							      echo "<td><a href='cloud.php?public=".basename($file)."'>Öffentlich verfügbar machen</a></td>";
+							      echo '</tr>';
+							}
+						}else{
+							echo '<tr>';
+							echo '<td><img  style="display:block; width:100px;height:100px;" id="base64image" src="data:image;base64,' . get_base64_preview($file) . '"/></td>';
+							echo '<td>' . basename($file) . '</td>';
+							echo '<td><a href="print.php?cloudprint='.basename($file).'">Drucken</a></td>';
+							echo "<td><a href='cloud.php?delete=".basename($file)."' >" . "Löschen" . '</a></td>';
+							echo "<td><a href='/system0/html/user_files/$username/".basename($file)."' download>" . "Herunterladen" . '</a></td>';
+							echo "<td><a href='cloud.php?public=".basename($file)."'>Öffentlich verfügbar machen</a></td>";
+							echo '</tr>';
+						}
 					  }
 				      } else {
 					  echo '<tr><td colspan="2">Directory not found</td></tr>';
